% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicPipes.R
\name{basic_backward}
\alias{basic_backward}
\alias{\%<\%}
\title{Basic Backward Pipe}
\usage{
basic_backward(lhs, rhs)
}
\arguments{
\item{lhs}{A function to apply to \code{rhs}.}

\item{rhs}{An object to be passed to the function.}
}
\value{
The result of applying \code{lhs} to \code{rhs}.
}
\description{
Passes the result of the right-hand side expression as the first argument to the left-hand side function.
}
