% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignmentPipes.R
\name{compound_assignment}
\alias{compound_assignment}
\alias{\%<>\%}
\title{Compound Assignment Pipe}
\usage{
compound_assignment(lhs, rhs)
}
\arguments{
\item{lhs}{A variable name.}

\item{rhs}{A function to apply to `lhs`.}
}
\value{
The result assigned back to `lhs` in the global environment.
}
\description{
Updates a variable by piping it through a function and reassigning the result.
}
