% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPipes.R
\name{dot_backward}
\alias{dot_backward}
\alias{\%<.\%}
\title{Dot Backward Pipe}
\usage{
dot_backward(lhs, rhs)
}
\arguments{
\item{lhs}{An expression using \code{.}.}

\item{rhs}{An object to assign to \code{.}.}
}
\value{
The result of evaluating \code{lhs} with \code{.} set to \code{rhs}.
}
\description{
Passes the right-hand value as \code{.} into the left-hand expression.
}
