% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPipes.R
\name{dot_forward}
\alias{dot_forward}
\alias{\%.>\%}
\title{Dot Forward Pipe}
\usage{
dot_forward(lhs, rhs)
}
\arguments{
\item{lhs}{An object to assign to \code{.}.}

\item{rhs}{An expression using \code{.}.}
}
\value{
The result of evaluating \code{rhs} with \code{.} set to \code{lhs}.
}
\description{
Passes the left-hand value as \code{.} into the right-hand expression.
}
