% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPipes.R
\name{dot_foward}
\alias{dot_foward}
\title{Dot Forward Pipe}
\usage{
dot_foward(lhs, rhs)
}
\arguments{
\item{lhs}{An object to assign to `.`.}

\item{rhs}{An expression using `.`.}
}
\value{
The result of evaluating `rhs` with `.` set to `lhs`.
}
\description{
Passes the left-hand value as `.` into the right-hand expression.
}
