% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expositionPipes.R
\name{exposition_backward}
\alias{exposition_backward}
\alias{\%<$\%}
\title{Exposition Backward Pipe}
\usage{
exposition_backward(lhs, rhs)
}
\arguments{
\item{lhs}{An expression to evaluate within \code{rhs}.}

\item{rhs}{A data object (e.g., a data frame).}
}
\value{
The result of evaluating \code{lhs} within \code{rhs}.
}
\description{
Evaluates a data object within the environment of an expression.
}
