% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expositionPipes.R
\name{exposition_forward}
\alias{exposition_forward}
\title{Exposition Forward Pipe}
\usage{
exposition_forward(lhs, rhs)
}
\arguments{
\item{lhs}{A data object (e.g., a data frame).}

\item{rhs}{An expression to evaluate within `lhs`.}
}
\value{
The result of evaluating `rhs` within `lhs`.
}
\description{
Evaluates an expression within the environment of a data object.
}
