% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teePipes.R
\name{tee_forward}
\alias{tee_forward}
\title{Tee Forward Pipe}
\usage{
tee_forward(lhs, rhs)
}
\arguments{
\item{lhs}{An object to be passed to `rhs`.}

\item{rhs}{A function to apply to `lhs` for side effects.}
}
\value{
Invisibly returns `lhs`.
}
\description{
Evaluates a function on the left-hand value for side effects, but returns the original left-hand value.
}
