% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throughPipes.R
\name{through_forward}
\alias{through_forward}
\title{Through Forward Pipe}
\usage{
through_forward(lhs, rhs)
}
\arguments{
\item{lhs}{An object to assign to `.`.}

\item{rhs}{An expression using `.`.}
}
\value{
Invisibly returns `lhs` after printing the result of `rhs`.
}
\description{
Evaluates an expression using `.` and returns the original value invisibly after printing the result.
}
