% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyePipes.R
\name{wye_backward}
\alias{wye_backward}
\alias{\%<Y\%}
\title{Wye Backward Pipe}
\usage{
wye_backward(lhs, rhs)
}
\arguments{
\item{lhs}{An expression evaluating to a list of functions.}

\item{rhs}{An expression evaluating to a single object.}
}
\value{
A vector of results from applying each function in `lhs` to `rhs`.
}
\description{
Applies a list of functions to a single input and returns a vector of results.
}
