% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyePipes.R
\name{wye_forward}
\alias{wye_forward}
\title{Wye Forward Pipe}
\usage{
wye_forward(lhs, rhs)
}
\arguments{
\item{lhs}{An expression evaluating to a single object.}

\item{rhs}{An expression evaluating to a list of functions.}
}
\value{
A vector of results from applying each function in `rhs` to `lhs`.
}
\description{
Applies a list of functions to a single input and returns a vector of results.
}
