% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotPipes.R
\name{dot_backward}
\alias{dot_backward}
\title{Dot Backward Pipe}
\usage{
dot_backward(lhs, rhs)
}
\arguments{
\item{lhs}{An expression using `.`.}

\item{rhs}{An object to assign to `.`.}
}
\value{
The result of evaluating `lhs` with `.` set to `rhs`.
}
\description{
Passes the right-hand value as `.` into the left-hand expression.
}
