% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throughPipes.R
\name{through_forward}
\alias{through_forward}
\alias{\%.>>\%}
\title{Through Forward Pipe}
\usage{
through_forward(lhs, rhs)
}
\arguments{
\item{lhs}{An object to assign to \code{.}.}

\item{rhs}{An expression using \code{.}.}
}
\value{
Invisibly returns \code{lhs} after printing the result of \code{rhs}.
}
\description{
Evaluates an expression using \code{.} and returns the original value invisibly after printing the result.
}
